requirements.txt
Django==4.2.16
django-cors-headers==4.7.0
djangorestframework==3.16.0
numpy==2.2.5
